backend/requirement.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
groq
psycopg2-binary 